requirements.txt
requests
pandas
openpyxl
python-docx
reportlab
urllib3
markdown